@@ -11,52 +11,51 @@ dependencies:
     sdk: flutter
 
   # State Management
-  provider: ^6.1.1
+  provider: ^6.1.5
   
   # HTTP Client
-  http: ^1.1.0
-  dio: ^5.4.0
+  http: ^1.6.0
+  dio: ^5.9.0
   
   # Local Storage
-  shared_preferences: ^2.2.2
+  shared_preferences: ^2.5.4
   
   # JSON Serialization
-  json_annotation: ^4.8.1
+  json_annotation: ^4.9.0
   
   # UI Components
-  cupertino_icons: ^1.0.6
-  flutter_svg: ^2.0.9
+  cupertino_icons: ^1.0.8
+  flutter_svg: ^2.2.3
   
   # Utilities
-  intl: ^0.18.1
-  connectivity_plus: ^5.0.2
+  intl: ^0.20.2
+  connectivity_plus: ^7.0.0
   
   # Camera & Image Processing
-  camera: ^0.10.5+5
-  image_picker: ^1.0.7
-  image: ^4.1.3
-  path_provider: ^2.1.1
+  camera: ^0.11.3
+  image_picker: ^1.2.1
+  image: ^4.7.2
+  path_provider: ^2.1.5
   
   # Location
-  geolocator: ^10.1.0
-  permission_handler: ^11.1.0
+  geolocator: ^14.0.2
+  permission_handler: ^12.0.1
   
   # Hashing
-  crypto: ^3.0.3
-  imagehash: ^1.0.0
+  crypto: ^3.0.7
   
   # File Upload
-  file_picker: ^6.1.1
+  file_picker: ^10.3.8
   
   # Maps (optional)
-  google_maps_flutter: ^2.5.0
+  google_maps_flutter: ^2.14.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^3.0.1
-  build_runner: ^2.4.7
-  json_serializable: ^6.7.1
+  flutter_lints: ^6.0.0
+  build_runner: ^2.10.4
+  json_serializable: ^6.11.3
 
 flutter:
   uses-material-design: true
