@@ -39,7 +39,7 @@ class AppTheme {
         filled: true,
         fillColor: Colors.white,
       ),
-      cardTheme: CardTheme(
+      cardTheme: CardThemeData(
         elevation: 2,
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(12),
@@ -79,7 +79,7 @@ class AppTheme {
         filled: true,
         fillColor: const Color(0xFF1E1E1E),
       ),
-      cardTheme: CardTheme(
+      cardTheme: CardThemeData(
         elevation: 2,
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(12),
